+++ src/main/java/roomescape/payment/service/PaymentService.java
@@ -5,8 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.exception.ErrorType;
 import roomescape.exception.NotFoundException;
 import roomescape.global.util.Encoder;
-import roomescape.payment.PaymentRequest;
-import roomescape.payment.PaymentResponse;
+import roomescape.payment.service.dto.PaymentRequest;
+import roomescape.payment.service.dto.PaymentResponse;
 import roomescape.payment.TossPaymentProperties;
 import roomescape.payment.domain.Payment;
 import roomescape.payment.domain.repository.PaymentRepository;
