+++ src/main/java/roomescape/payment/service/dto/PaymentErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.payment.service.dto;
 
 public record PaymentErrorResponse(String code, String message) {
 }
