+++ src/main/java/roomescape/payment/service/dto/PaymentRequest.java
@@ -1,4 +1,4 @@
-package roomescape.payment;
+package roomescape.payment.service.dto;
 
 public record PaymentRequest(
         long amount,
