+++ src/test/java/roomescape/payment/service/PaymentServiceTest.java
@@ -22,9 +22,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.repository.MemberRepository;
-import roomescape.payment.PaymentErrorResponse;
-import roomescape.payment.PaymentRequest;
-import roomescape.payment.PaymentResponse;
+import roomescape.payment.service.dto.PaymentErrorResponse;
+import roomescape.payment.service.dto.PaymentRequest;
+import roomescape.payment.service.dto.PaymentResponse;
 import roomescape.payment.domain.PayAmount;
 import roomescape.payment.domain.Payment;
 import roomescape.payment.domain.PaymentType;
