+++ src/test/java/roomescape/reservation/service/ReservationApplicationServiceTest.java
@@ -27,8 +27,8 @@ import roomescape.exception.BadRequestException;
 import roomescape.exception.ErrorType;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.repository.MemberRepository;
-import roomescape.payment.PaymentRequest;
-import roomescape.payment.PaymentResponse;
+import roomescape.payment.service.dto.PaymentRequest;
+import roomescape.payment.service.dto.PaymentResponse;
 import roomescape.reservation.controller.dto.ReservationResponse;
 import roomescape.reservation.domain.MemberReservation;
 import roomescape.reservation.domain.Reservation;
